+++ src/index.ts
@@ -1,6 +1,4 @@
 #!/usr/bin/env node
-// This must me converted into a dynamic import
-import watchAndBuild from './watchAndBuild'
 import getOptions from './getOptions'
 
 const options = getOptions()
@@ -8,16 +6,15 @@ const options = getOptions()
 const main = async () => {
   if (options.version) {
     const { displayVersion } = await import('./displayVersion')
-    displayVersion()
-  } else if (options.watch) {
-    const chokidar = await import('chokidar')
-    chokidar
-      .watch(options.watch)
-      .on('ready', async () => watchAndBuild())
-      .on('change', async () => watchAndBuild())
-  } else {
-    watchAndBuild()
+    return displayVersion()
   }
+  const { watchAndBuild } = await import('./watchAndBuild')
+  options.watch
+    ? (await import('chokidar'))
+        .watch(options.watch)
+        .on('ready', async () => watchAndBuild())
+        .on('change', async () => watchAndBuild())
+    : watchAndBuild()
 }
 
 main()